Use uint64_t in software too.

+++ VivadoTest/Software.cpp
@@ -2,7 +2,7 @@
 #include "Software.h"
 #include "Gonzo.h"
 
-void testGonzoSoftware(gonzo::TriangleArray triangles, int numTriangles, gonzo::VertexArray vertices, gonzo::BBox3fa& box) {
+void testGonzoSoftware(gonzo::TriangleArray triangles, uint64_t numTriangles, gonzo::VertexArray vertices, gonzo::BBox3fa& box) {
 
 	gonzo::Gonzo gonzo;
 	gonzo.build(triangles, numTriangles, vertices, box);
