+++ VivadoTest/Software.h
@@ -4,4 +4,4 @@
 #include "TriangleMesh.h"
 #include "vec3.h"
 
-void testGonzoSoftware(gonzo::TriangleArray triangles, int numTriangles, gonzo::VertexArray vertices, gonzo::BBox3fa& box);
+void testGonzoSoftware(gonzo::TriangleArray triangles, uint64_t numTriangles, gonzo::VertexArray vertices, gonzo::BBox3fa& box);
